Fix erroneous dag circuit depth less than zero
Method `depth` was miscopied (translated) from the Python implementation.
The latter returns zero if the longest path length is zero rather
than the longest path length minus one.

In a circuit with zero qubits the longest path is of length zero.
If this were not handled specially, a depth of -1 would be returned.

The fix also uses checked subtraction. Before, the underflow of returning
type usize and value -1 caused a panic.